rename Node to QuadTreeNode
