+++ .settings/org.eclipse.jdt.ui.prefs
@@ -1,0 +1,3 @@
+eclipse.preferences.version=1
+formatter_profile=_GoogleStyle
+formatter_settings_version=12
